requirements.txt
flask==2.3.2
twilio==8.5.0
python-dotenv==0.15.0
openai==0.27.8
gunicorn==21.2.0
celery==5.3.1
redis==4.5.4
